backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mini-dpp-platform"
version = "0.1.0"
description = "Mini Digital Product Passport Platform based on AAS and IDTA DPP4.0"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "DPP Platform Team" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
]

dependencies = [
    # Web Framework
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",

    # Database
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "greenlet>=3.0.0",

    # Caching
    "redis>=5.0.0",
    "hiredis>=2.3.0",

    # Authentication & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.26.0",
    "cryptography>=42.0.0",

    # AAS Processing
    "xmltodict>=0.13.0",
    "jsonschema>=4.21.0",
    "basyx-python-sdk==2.0.0",
    "rdflib>=7.0.0",

    # Validation & Serialization
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "orjson>=3.9.0",

    # QR Code Generation
    "qrcode[pil]>=7.4.0",
    "pillow>=10.2.0",

    # Object Storage
    "minio>=7.2.0",

    # PDF Export
    "fpdf2>=2.7.8",

    # Cryptographic Audit Trail
    "asn1crypto>=1.5.0",

    # Metrics
    "prometheus-fastapi-instrumentator>=7.0.0",

    # YAML processing
    "pyyaml>=6.0.0",

    # Utilities
    "python-dateutil>=2.8.2",
    "tenacity>=8.2.0",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "httpx>=0.26.0",
    "factory-boy>=3.3.0",
    "faker>=22.0.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "testcontainers>=3.7.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C4",
    "UP",
    "ARG",
    "SIM",
]
ignore = ["E501", "B008", "B904"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]
exclude = ["app/db/migrations/"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.bandit]
targets = ["app"]
exclude_dirs = ["tests"]
skips = ["B101"]  # Allow assert in non-test code (used by Pydantic validators)

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "types-defusedxml>=0.7.0.20250822",
]
